refactor(LottoWinningTicket): LottoWinningNumber.java -> LottoWinningTicket 으로 이름 변경 refactor(package): 패키지 구조 변경
