+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.ticket.LottoTicket;
 import lotto.domain.vo.LottoMoney;
 
 import java.util.ArrayList;
