+++ src/test/java/lotto/domain/number/LottoWinningTicketTest.java
@@ -1,5 +1,6 @@
 package lotto.domain.number;
 
+import lotto.domain.ticket.LottoWinningTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,9 +9,8 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
-class LottoWinningNumberTest {
+class LottoWinningTicketTest {
 
     @DisplayName("당첨 번호와 보너스 번호가 중복 되는 것이 있으면 IllegalArgumentException throw")
     @Test
@@ -19,7 +19,7 @@ class LottoWinningNumberTest {
         int bonusNumber = 5;
 
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> new LottoWinningNumber(winningNumbers, bonusNumber))
+                .isThrownBy(() -> new LottoWinningTicket(winningNumbers, bonusNumber))
                 .withMessage("보너스 번호는 당첨 번호와 중복 될 수 없습니다. - " + bonusNumber);
     }
 
@@ -30,7 +30,7 @@ class LottoWinningNumberTest {
         int bonusNumber = 8;
 
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> new LottoWinningNumber(winningNumbers, bonusNumber))
+                .isThrownBy(() -> new LottoWinningTicket(winningNumbers, bonusNumber))
                 .withMessage("로또 번호는 6개만 가능합니다. - " + winningNumbers);
     }
 
@@ -41,7 +41,7 @@ class LottoWinningNumberTest {
         int bonusNumber = 8;
 
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> new LottoWinningNumber(winningNumbers, bonusNumber))
+                .isThrownBy(() -> new LottoWinningTicket(winningNumbers, bonusNumber))
                 .withMessageContaining("로또 번호는 중복 될 수 없습니다.");
     }
 }
